requirements.txt
baal==2.0.0
colorama==0.4.6
environs==9.5.0
# faiss==1.7.4
flatdict==4.0.1
# jax==0.4.34
lightning==2.2.5
loss_landscapes==3.0.6
matplotlib==3.8.0
netcal==1.1.2
neural_tangents==0.6.5
numpy==2.1.2
omegaconf==2.3.0
pandas==2.2.3
Pillow==10.0.1
Pillow==10.4.0
pytest==7.4.0
scikit_learn==1.4.2
scikit_learn==1.3.2
scipy==1.14.1
skimage==0.0
statsmodels==0.14.1
structlog==24.4.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.65.0
tqdm==4.66.1
typing_extensions==4.12.2
Wand==0.6.10
wandb==0.15.12
wget==3.2
wilds==2.0.0
